+++ src/bin/server/main.rs
@@ -1,66 +1,116 @@
-use std::io::{Read, Write};
-use std::net::TcpListener;
+
+use std::collections::HashMap;
+use std::net::{SocketAddr, TcpListener, TcpStream};
 use std::sync::{Arc, Mutex};
-use std::thread;
-use std::str;
+
 
 fn main() {
-    let listener = TcpListener::bind("0.0.0.0:2925").expect("Could not bind");
-    let running = Arc::new(Mutex::new(true));
+        
+}
 
-    println!("Server is listening on port 2925");
+pub type PlayerCollection = HashMap<SocketAddr, TcpStream>;
 
-    let running_clone = Arc::clone(&running);
-    thread::spawn(move || {
-        // This thread listens for the termination command
-        let mut buffer = String::new();
-        std::io::stdin().read_line(&mut buffer).expect("Failed to read line");
-        let mut running = running_clone.lock().unwrap();
-        *running = false;
-    });
+#[derive(Debug, PartialEq)]
+pub enum ServerStatus {
+    WaitingForHost,
+    WaitingForPlayers,
+    Busy,
+    Idle,
+    Active,
+    Inactive,
+}
 
-    for stream in listener.incoming() {
+pub struct GameServer {
+    status: Arc<Mutex<ServerStatus>>,
+    players: PlayerCollection,
+    number_of_players: u16,
+    port: u16,
+}
+
+impl GameServer {
+    pub fn start(status: Arc<Mutex<ServerStatus>>, number_of_players: u16, port: u16) {
+        
+        let mut players: PlayerCollection = wait_for_players(number_of_players, port, &status);
+        for (addr, _) in players.iter() {
+            println!("[PLAYER] {} is ready!", addr);
+        }
+        
         {
-            let running = running.lock().unwrap();
-            if !*running {
-                println!("Server is shutting down");
-                break;
-            }
+            let mut server_status_lock = status.lock().unwrap();
+            *server_status_lock = ServerStatus::WaitingForHost;
+            drop(server_status_lock);
         }
 
-        match stream {
-            Ok(mut stream) => {
-                println!("Client connected");
-                let mut buffer = [0; 512];
-                loop {
-                    match stream.read(&mut buffer) {
-                        Ok(bytes_read) => {
-                            if bytes_read == 0 {
-                                println!("Client disconnected");
-                                break;
-                            }
+        let mut game_server: GameServer = GameServer {
+            status,
+            players: HashMap::new(),
+            number_of_players,
+            port,
+        };
 
-                            match str::from_utf8(&buffer[..bytes_read]) {
-                                Ok(client_message) => {
-                                    println!("Client says: {}", client_message);
-                                }
-                                Err(_) => {
-                                    println!("Received invalid UTF-8 message: {:?}", &buffer[..bytes_read]);
-                                }
-                            }
+        game_server.run();
+    }
 
-                            stream.write_all(b"Trail Server > ").expect("Could not write to stream");
-                        }
-                        Err(e) => {
-                            println!("Failed to read from stream: {}", e);
-                            break;
-                        }
+    fn run(&mut self) {
+        // game_loop(&self.number_of_players, &mut self.players);
+    }
+}
+
+
+
+pub fn wait_for_players(number_of_players: u16, port: u16, server_status: &Arc<Mutex<ServerStatus>>) -> PlayerCollection {
+    let listener: TcpListener = TcpListener::bind(("0.0.0.0", port)).expect("Could not bind to port");
+
+    {
+        let mut server_status_lock = server_status.lock().unwrap();
+        *server_status_lock = ServerStatus::WaitingForHost;
+        drop(server_status_lock);
+    }
+
+    let mut players: PlayerCollection = HashMap::new();
+    let mut host_joined: bool = false;
+
+    println!(
+        "[SERVER] Waiting for {} players to connect on port {}...",
+        number_of_players, port
+    );
+
+    // TODO: May need something about making sure the first connection comes from 127.0.0.1
+    while players.len() < number_of_players as usize {
+        
+        match listener.accept() {
+
+            Ok((stream, addr)) => {
+                
+                if !host_joined && addr.ip().to_string() == "127.0.0.1"{
+                    println!("[HOST CONNECTED] {}", addr);
+                    host_joined = true;
+                    {
+                        let mut server_status_lock = server_status.lock().unwrap();
+                        *server_status_lock = ServerStatus::WaitingForPlayers;
+                        drop(server_status_lock);
                     }
+
+                } else if host_joined {
+                    println!("[PLAYER CONNECTED] {}", addr);
+                    players.insert(addr, stream);
+                    println!(
+                        "[PLAYERS CONNECTED] {}/{}",
+                        &players.len(),
+                        number_of_players
+                    );
+
+                } else {
+                    println!("[ERROR] Host must connect first.");
                 }
             }
             Err(e) => {
-                println!("Connection failed: {}", e);
+                println!("[ERROR] Could not accept connection: {}", e);
             }
         }
     }
+
+    println!("[SERVER] All players connected.");
+
+    players
 }
